[feat] Update update.sh: add --peek, better messaging

+++ update.sh
@@ -5,8 +5,10 @@
 if [ "$1" == "-h" ] || [ "$1" == "--help" ]; then
   echo "Fetch and update Abseil GN build files"
   echo ""
+  echo "Alterative behaviors:"
   echo "  -h, --help  show this help message and exit"
   echo "  -r, --reuse skip downloading from upstream"
+  echo "  -p, --peek  check if a new Abseil revision is available"
   exit 0
 fi
 
@@ -17,6 +19,20 @@ abseil_rev=$dirname/ABSL_REVISION.txt
 abseil_url=https://github.com/abseil/abseil-cpp.git
 
 set -e # Exit on error
+
+if [ "$1" = "-p" ] || [ "$1" = "--peek" ]; then
+  local_commit=$(tail -1 $abseil_rev)
+  echo "local commit:  $local_commit"
+  printf "remote commit: "
+  remote_commit=$(git ls-remote $abseil_url HEAD | cut -f1)
+  printf "\rremote commit: $remote_commit\n"
+  if [ $local_commit = $remote_commit ]; then
+    echo "no update"
+  else
+    echo "new revision available"
+  fi
+  exit 0
+fi
 
 if [ "$1" != "-r" ] && [ "$1" != "--reuse" ]; then
   rm -rf absl
@@ -34,4 +50,7 @@ fi
 $dirname/gen.py . --profile=$dirname/absl.json # Help: use "--help"
 
 echo "Rolled to Abseil commit: "
-cat ABSL_REVISION.txt
+cat $abseil_rev
+
+commit_hash=$(tail -1 $abseil_rev | cut -c 1-7)
+echo "Do: git add . && git commit -m \"[roll] Roll to Abseil $commit_hash\""
